+++ lib/rubocop/cop/style/extra_spacing.rb
@@ -1,5 +1,7 @@
 # encoding: utf-8
 
+require 'set'
+
 module RuboCop
   module Cop
     module Style
@@ -21,23 +23,27 @@ module RuboCop
         include PrecedingFollowingAlignment
 
         MSG_UNNECESSARY = 'Unnecessary spacing detected.'
-        MSG_UNALIGNED_ASGN = '`=` is not aligned with the previous assignment.'
+        MSG_UNALIGNED_ASGN = '`=` is not aligned with the %s assignment.'
 
         def investigate(processed_source)
           ast = processed_source.ast
 
-          asgn_lines = processed_source
-                       .tokens
-                       .select { |t| equal_sign?(t) }
-                       .map { |t| t.pos.line }
+          if force_equal_sign_alignment?
+            @asgn_tokens = processed_source.tokens.select { |t| equal_sign?(t) }
+            # Only attempt to align the first = on each line
+            @asgn_tokens = Set.new(@asgn_tokens.uniq { |t| t.pos.line })
+            @asgn_lines  = @asgn_tokens.map { |t| t.pos.line }
+            # Don't attempt to correct the same = more than once
+            @corrected   = Set.new
+          end
 
           processed_source.tokens.each_cons(2) do |t1, t2|
             next if t2.type == :tNL
-            next if t1.pos.line != t2.pos.line
 
             if force_equal_sign_alignment? &&
-               equal_sign?(t2) &&
-               asgn_lines.include?(t2.pos.line - 1)
+               @asgn_tokens.include?(t2) &&
+               (@asgn_lines.include?(t2.pos.line - 1) ||
+                @asgn_lines.include?(t2.pos.line + 1))
               check_assignment(t2)
             else
               check_other(t1, t2, ast)
@@ -45,8 +51,19 @@ module RuboCop
           end
         end
 
-        def autocorrect(range)
-          ->(corrector) { corrector.remove(range) }
+        def correct(range)
+          return :uncorrected unless autocorrect?
+
+          if range.source.end_with?('=')
+            align_equal_sign(range)
+          else
+            correction { |corrector| corrector.remove(range) }
+          end
+          :corrected
+        end
+
+        def support_autocorrect?
+          true
         end
 
         private
@@ -55,10 +72,15 @@ module RuboCop
           # minus 2 is because pos.line is zero-based
           line = processed_source.lines[token.pos.line - 2]
           return if aligned_assignment?(token.pos, line)
-          add_offense(token.pos, token.pos, MSG_UNALIGNED_ASGN)
+
+          preceding  = @asgn_lines.include?(token.pos.line - 1)
+          align_with = preceding ? 'preceding' : 'following'
+          message    = format(MSG_UNALIGNED_ASGN, align_with)
+          add_offense(token.pos, token.pos, message)
         end
 
         def check_other(t1, t2, ast)
+          return if t1.pos.line != t2.pos.line
           return if t2.pos.begin_pos - 1 <= t1.pos.end_pos
           return if allow_for_alignment? && aligned_token?(t2)
 
@@ -124,6 +146,51 @@ module RuboCop
 
         def equal_sign?(token)
           token.type == :tEQL || token.type == :tOP_ASGN
+        end
+
+        def align_equal_sign(range)
+          lines  = contiguous_assignment_lines(range)
+          tokens = @asgn_tokens.select { |t| lines.include?(t.pos.line) }
+
+          columns  = tokens.map { |t| align_column(t) }
+          align_to = columns.max
+
+          tokens.each do |token|
+            next unless @corrected.add?(token)
+            diff = align_to - token.pos.last_column
+
+            if diff > 0
+              correction { |corr| corr.insert_before(token.pos, ' ' * diff) }
+            elsif diff < 0
+              correction { |corr| corr.remove_preceding(token.pos, -diff) }
+            end
+          end
+        end
+
+        def contiguous_assignment_lines(range)
+          result = [range.line]
+
+          range.line.downto(1) do |lineno|
+            @asgn_lines.include?(lineno) ? result << lineno : break
+          end
+          range.line.upto(processed_source.lines.size) do |lineno|
+            @asgn_lines.include?(lineno) ? result << lineno : break
+          end
+
+          result.sort!
+        end
+
+        def align_column(asgn_token)
+          # if we removed unneeded spaces from the beginning of this =,
+          # what column would it end from?
+          line    = processed_source.lines[asgn_token.pos.line - 1]
+          leading = line[0...asgn_token.pos.column]
+          spaces  = leading.size - (leading =~ / *\Z/)
+          asgn_token.pos.last_column - spaces + 1
+        end
+
+        def correction(&block)
+          @corrections << block
         end
       end
     end